Refactor Waveform generation, rescaling
Move generating, rescaling into separate units and separate threads
